move distibution steps to nw component 1. DistributionSteps livewire component 2. distribution-steps livewire component 3. integrate diistribution-steps with add-distribution view component
