+++ resources/views/components/data-table/table.blade.php
@@ -1,5 +1,5 @@
 @aware(['entries'=>true])
-<table {{ $attributes->merge(['class' => 'min-w-full '])}}>
+<table {{ $attributes->merge(['class' => 'w-full'])}}>
     @if ($entries)
     <x-data-table.entries />
     @endif
