+++ resources/views/components/form/table.blade.php
@@ -7,13 +7,13 @@ else {
 $search = "";
 }
 @endphp
-<div class="relative py-6 px-5 sm:px-6 lg:py-0 lg:px-0 lg:col-span-8 bg-white sm:rounded-lg">
+<div class="relative py-6 px-5 sm:px-6 lg:py-3 lg:px-0 bg-white sm:rounded-lg overflow-auto">
 
     <x-form.header search="{{ $search }}">
         {{ __($title) }}
     </x-form.header>
 
-    <x-data-table.table class="myTable2 mb-14 overflow-x-auto">
+    <x-data-table.table class="myTable2 mb-14">
         <thead class="bg-lime-50">
             <x-data-table.tr>
                 {{ $tableHeaders ?? $slot }}
